Add handling for when runner.test_start_time is nil
When running in an older version of minitest, our selenium
integration (based on MiniTest::Spec) does not call #before_test
before calling #print_test_with_time for the first time.
Defaulting test_start_time to Time.now does not seem to effect
the rest of the test run, and times are reported correctly.