Add 'MeanTimeReporter';
Using a temporary 'previous runs' file, we gather statistics on previous
test run times. From this file we build a report which indicates the average
test run time, a minimum (fastest) test run time, a maximum (slowest) test
run time for each test. A description is included for each report line which
is the test name as reported by Minitest.