Fix a bug that RE-generated JUnit filenames differently
If two describe blocks have the same description, JUnit creates the same
filename for them. To avoid overwriting reports, it adds a number to the
filename if it finds a file already existing.

Before this change, the filename with the number was generated
differently than the original filename-- namely, it did not include the
special character substitution or filename length limiting. Therefore,
if two describes were the same and both used special characters, the
second one would fail because its filename would be invalid.

To fix this, reuse the name generated without the special characters and
limited to a filesystem-friendly length, even when we're adding a number
to avoid filename collisions.