Implement DelegateReporter#io
This makes the duck type of DelegateReporter more similar to
Minitest::StatisticsReporter avoiding a crash if #io is called but not
delegated to one of the reporters. If the io option was passed to
DelegateReporter.new all of the reporters' io attributes should be
initialized to it anyway, so just return it.