Use block syntax to close files after writing
After adding minitest-reporters to two large Rails apps in order to use
the MeanTimeReporter to identify the slowest running tests.

I started seeing `Errno::EMFILE: Too many open files` errors. The errors
occour intermitantly locally, but consistently on CI. I don't really
think that it's minitest-reporters "fault", however when reviewing the
code I noticed that there were a couple places in the reporter that file
handles could be being left open.

This PR switches to using the block syntax to open and then write to a
file (as is done elsewhere in the reporter), which implicitly closes the
file when exiting the block.