Fix: Last test suite is Minitest::Result
The last suite to be evaluated here is a Minitest::Result and not an
instance of the test class. This results in the time reporters reporting
the last suite as "Minitest::Result" instead of the suite which was
actually run.

We can use the #test_helper already defined to parse this for us