Remove debug output from ProgressReporter
I noticed that the ProgressReporter was debugging details about failed tests twice: one time as  inspected array, then as proper string. Also, the class name and test name were reversed. I suspect the inspected array was a debug statement that accidentally was left in place.

I removed the debug print statement, and fixed the order of the test class and method name to be `Suite#test_method`.