Add alias for method name with misspelling in HTML reporter
Previously, `HtmlReporter` defined a method `percent_skipps`. This
was a misspelling that has been corrected to `percent_skips`. However,
it's possible some users of minitest-reporters already created custom
HTML/ERB templates depending on the misspelling.

This commit defines a method alias for the misspelled version of the
method to avoid breaking any existing custom templates.

https://github.com/minitest-reporters/minitest-reporters/pull/326