Switch progress bar default format to proper fraction
This patch switches the progress bar's default format to display the
completed test count in the numerator and the total test count in the
denominator, as in "52/2182" rather than "2182/52". Displaying progress
as a proper fraction is more intuitive given the fraction-like notation
we use here and corresponds nicely to the "2%" completion status that we
also display in the default format.

The bar's format is configurable, so users who prefer seeing the old
"total/completed" format should instantiate their reporter as follows:

```ruby
Minitest::Reporters::ProgressReporter.new(format: '  %C/%c: [%B] %p%% %a, %e')
```